+++ modules/flann/include/opencv2/flann/result_set.h
@@ -299,7 +299,7 @@ public:
 	{
 		Item it;
 		it.index = index;
-		it.dist = flann_dist(target, target_end, point);
+		it.dist = (float)flann_dist(target, target_end, point);
 		if (it.dist<=radius) {
 			items.push_back(it);
 			push_heap(items.begin(), items.end());
